+++ src/infra/database/postgres/migrations/1707306520428-create-food.ts
@@ -6,10 +6,10 @@ export class CreateFood1707306520428 implements MigrationInterface {
             CREATE TABLE "food" (
                 "id" UUID NOT NULL DEFAULT uuid_generate_v4(),
                 "name" VARCHAR(255) NOT NULL,
-                "calories" INT NOT NULL,
-                "fat" INT NOT NULL,
-                "carbohydrates" INT NOT NULL,
-                "protein" INT NOT NULL,
+                "calories" FLOAT NOT NULL,
+                "fat" FLOAT NOT NULL,
+                "carbohydrates" FLOAT NOT NULL,
+                "protein" FLOAT NOT NULL,
                 "image_url" VARCHAR(255),
                 CONSTRAINT "PK_food_id" PRIMARY KEY ("id")
             );
